backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.7
chromadb==0.4.15
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
numpy==1.24.3
tiktoken==0.5.2